+++ app/src/main/java/com/example/weatherapp/home/data/source/remote/RemoteDataSource.kt
@@ -4,6 +4,6 @@ import com.example.weatherapp.home.data.source.remote.model.CityWeatherResponse
 import io.reactivex.Single
 
 interface RemoteDataSource {
-    fun getCityWeatherByLatLon(lat: Long, lon: Long): Single<CityWeatherResponse>
+    fun getCityWeatherByLatLon(lat: Double, lon: Double): Single<CityWeatherResponse>
     fun getCityWeatherByCityName(cityName: String): Single<CityWeatherResponse>
 }
