+++ app/src/main/java/com/example/weatherapp/home/domain/interactor/GetCityWeatherByLatLongUseCase.kt
@@ -4,5 +4,5 @@ import com.example.weatherapp.home.domain.repository.CityWeatherRepository
 import javax.inject.Inject
 
 class GetCityWeatherByLatLongUseCase @Inject constructor(private val cityWeatherRepository: CityWeatherRepository) {
-    fun execute(lat: Long, long: Long) = cityWeatherRepository.getCityWeatherByLatLong(lat, long)
+    fun execute(lat: Double, long: Double) = cityWeatherRepository.getCityWeatherByLatLong(lat, long)
 }
